Task 6. Size is not grandeur, and territory does not make a nation

+++ 0x00-hello_world/6-size.c
@@ -15,10 +15,10 @@ int main(void)
 	long long lli;
 	float f;
 
-	printf("Size of a char: %d byte(s)\n", sizeof(c));
-	printf("Size of an int: %d byte(s)\n", sizeof(i));
-	printf("Size of a long int: %d byte(s)\n", sizeof(li));
-	printf("Size of a long long int: %d byte(s)\n", sizeof(lli));
-	printf("Size of a float: %d byte(s)\n", sizeof(f));
+	printf("Size of a char: %d byte(s)\n", (int) sizeof(c));
+	printf("Size of an int: %d byte(s)\n", (int) sizeof(i));
+	printf("Size of a long int: %d byte(s)\n", (int) sizeof(li));
+	printf("Size of a long long int: %d byte(s)\n", (int) sizeof(lli));
+	printf("Size of a float: %d byte(s)\n", (int) sizeof(f));
 	return (0);
 }
